marketing-agent/pyproject.toml
[tool.poetry]
name = "marketing-agent"
version = "0.1.0"
description = ""
authors = [
    "Leo Kim <leo.kim@cerebras.net>",
    "Kunal Patel <kunal.patel@cerebras.net>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cerebras_cloud_sdk = "^1.0.0"
fireworks-ai = "^0.15.0"
groq = "^0.9.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
together = "^1.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
